add staggered button animation to something's the same trials
